chore: clear `task lint` (clippy + rustfmt) across the tree
`task lint` runs `cargo clippy --all --all-features --tests -- -D warnings`
and was failing with 24 errors. Fixed each one:

- **optimization_policy.rs**: derive `Default` on `OptimizationMode`
  (with `#[default] Auto`) instead of the manual impl; flatten the
  two `if ... { false } else { true }` Auto-branch arms into single
  expressions (`needless_bool`).
- **render.rs**: drop the `as u32` cast on `coverage_buffers.capacity`
  (already `u32`).
- **lib.rs**: separate the `COMPATIBITLIY_REQUIREMENTS` storage-buffer
  list from the following prose paragraph with a blank `///` line
  (`doc_lazy_continuation`).
- **light_buckets/buckets.rs**: same blank-line fix for the oversized-
  thresholds doc; also drop the never-called test-only `light_order`
  getter (`dead_code`).
- **meshes/meta.rs**: reword the `patch_mesh_light_slice` doc so it
  doesn't start a line with `+`, which markdown reads as a list
  marker (`doc_lazy_continuation`).
- **material_decal/classify/buffers.rs**: same — replace the line-
  leading `+` in `header_scratch`'s doc.
- **hzb/texture.rs, material_decal/composite.rs, occlusion/compaction.rs**:
  drop `useless_conversion` `.map_err(AwsmCoreError::from)` calls
  (the underlying `?` already coerces). Removed an unused
  `AwsmCoreError` import.
- **scene_spatial/tests.rs**: replace `|k| slotmap::Key::data(k)`
  with the bare method ref (`redundant_closure`).
- **scene-editor/measurement.rs**: use `?` on the
  `r.camera.last_matrices.as_ref()` shortcut (`question_mark`); also
  picks up the earlier camera_pos fix (was still using the broken
  `view.inverse().transpose().w_axis` formula that was corrected in
  `shadows::importance`, so the measurement histogram now reflects
  what the live scoring actually produces).
- **scene-editor/properties/kind_editor/decal.rs**: collapse the
  trivial single-arm `match select.value().as_str()` — only one
  blend mode is exposed today; drop the unreachable match scaffold
  (`match_single_binding`).

Plus rustfmt of the touched files. `task lint` exits 0; 114 lib
tests still pass.

+++ crates/renderer/src/render.rs
@@ -999,7 +999,7 @@ impl AwsmRenderer {
         // coverage rather than ringing the buffer).
         if let Some(slot_map) = coverage_slot_map {
             let readback_buffer = self.coverage_buffers.readback_buffer.clone();
-            let readback_size_bytes = (self.coverage_buffers.capacity as u32).saturating_mul(4);
+            let readback_size_bytes = self.coverage_buffers.capacity.saturating_mul(4);
             let state = self.coverage_readback_state.clone();
             state.borrow_mut().inflight = true;
             wasm_bindgen_futures::spawn_local(async move {
